+++ test/unit/test_api.py
@@ -5,8 +5,8 @@ from harvey.api import Api
 
 @patch('logging.Logger.info')
 @patch('harvey.pipelines.Pipeline.run_pipeline')
-def test_parse_webhook(mock_run_pipeline, mock_logger, mock_webhook_object):
-    webhook = Api.parse_webhook(mock_webhook_object)
+def test_parse_github_webhook(mock_run_pipeline, mock_logger, mock_webhook_object):
+    webhook = Api.parse_github_webhook(mock_webhook_object)
 
     mock_logger.assert_called()
     assert webhook[0] == {
@@ -18,8 +18,8 @@ def test_parse_webhook(mock_run_pipeline, mock_logger, mock_webhook_object):
 
 @patch('logging.Logger.debug')
 @patch('harvey.pipelines.Pipeline.run_pipeline')
-def test_parse_webhook_bad_branch(mock_run_pipeline, mock_logger, mock_webhook_object):
-    webhook = Api.parse_webhook(mock_webhook_object(branch='bad_branch_name'))
+def test_parse_github_webhook_bad_branch(mock_run_pipeline, mock_logger, mock_webhook_object):
+    webhook = Api.parse_github_webhook(mock_webhook_object(branch='bad_branch_name'))
 
     mock_logger.assert_called()
     assert webhook[0] == {
@@ -31,10 +31,10 @@ def test_parse_webhook_bad_branch(mock_run_pipeline, mock_logger, mock_webhook_o
 
 @patch('logging.Logger.debug')
 @patch('harvey.pipelines.Pipeline.run_pipeline')
-def test_parse_webhook_no_json(mock_run_pipeline, mock_logger):
+def test_parse_github_webhook_no_json(mock_run_pipeline, mock_logger):
     mock_webhook = MagicMock()
     mock_webhook.json = None
-    webhook = Api.parse_webhook(mock_webhook)
+    webhook = Api.parse_github_webhook(mock_webhook)
 
     mock_logger.assert_called()
     assert webhook[0] == {
@@ -48,8 +48,8 @@ def test_parse_webhook_no_json(mock_run_pipeline, mock_logger):
 @patch('harvey.config.Config.webhook_secret', '123')
 @patch('harvey.webhooks.Webhook.validate_webhook_secret', return_value=False)
 @patch('harvey.pipelines.Pipeline.run_pipeline')
-def test_parse_webhook_bad_webhook_secret(mock_run_pipeline, mock_logger, mock_webhook_object):
-    webhook = Api.parse_webhook(mock_webhook_object)
+def test_parse_github_webhook_bad_webhook_secret(mock_run_pipeline, mock_logger, mock_webhook_object):
+    webhook = Api.parse_github_webhook(mock_webhook_object)
 
     mock_logger.assert_called()
     assert webhook[0] == {