+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## NEXT RELEASE
 
+* Adds a simple authentication system where endpoints can be protected with the same API key as the webhook secret (uses basic auth)
 * Moved all `lock` functionality to its own module
 * Added new `/locks`, `/locks/{project_name}`, `/projects/{project_name}/lock`, and `/projects/{project_name}/unlock` endpoints
 * Fixed a bug where retrieving a lock status may return the wrong value